Fix/merge fix(#10) - 중복 및 변경된 코드를 하나의 코드로 일치시키기

+++ src/main/java/com/ddang/usedauction/point/domain/PointHistory.java
@@ -1,7 +1,7 @@
 package com.ddang.usedauction.point.domain;
 
 import com.ddang.usedauction.config.BaseTimeEntity;
-import com.ddang.usedauction.member.Member;
+import com.ddang.usedauction.member.domain.Member;
 import com.ddang.usedauction.point.type.PointType;
 import jakarta.persistence.*;
 import lombok.*;
@@ -19,17 +19,15 @@ public class PointHistory extends BaseTimeEntity {
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private PointType pointType;
+    private PointType pointType; // 포인트 타입
 
-    // 충전 or 사용 포인트량
     @Column(nullable = false)
-    private int pointAmount;
+    private long pointAmount; // 충전 or 사용 포인트량
 
-    // 현재 보유 포인트량
     @Column(nullable = false)
-    private int curPointAmount;
+    private long curPointAmount; // 현재 보유 포인트량
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
+    private Member member; // 회원
 }
